add notification for TBO expiration and update horasOperacaoMotor field type

+++ src/main/java/puc/airtrack/airtrack/OrdemDeServico/CabecalhoOrdemDTO.java
@@ -68,7 +68,7 @@ public class CabecalhoOrdemDTO {
     private String engenheiroAtuanteNome;
 
     @JsonProperty("horas_operacao_motor")
-    private Float horasOperacaoMotor;
+    private int horasOperacaoMotor;
 
     @JsonProperty("status_descricao")
     public String getStatusDescricao() {
